Merged PR 4094863: Fixing Java pipeline to ensure using Azul Java as it was expected
Java automatic build pipeline expected `%JAVA_HOME%` to not be set, so it can pick up the `%DEFAULT_JAVA_HOME%` (which uses Azul version, in sync with java compiler that are used in DS_Main_Dev TestShellTests in Extensibility).
Open-source extension, however, wants to let users set their own `%JAVA_HOME%` if they want.

To this end, in the pipeline we send an additional argument, denoted as `azuljava`, enforcing the pipeline to use `%DEFAULT_JAVA_HOME%`.  The arguments for open source expected to be debug/release/both, which even with this change still would work. Indeed, as long as users don't send the keyword `azuljava` as their first arg, they can have their own `JAVA_HOME` for compilation.