Merged PR 4895846: Speeding up IsValidUTF8 by changing the function signature + the local build by -j param
**Issue**
Test suite Extensibility.ThirdParty.RExtension.StreamInputDataQueryIsm no longer meets runtime for GREEN and has been moved to PROBATION.
The following work item was created for it automatically by cloudtest:
https://dev.azure.com/msdata/Database%20Systems/_workitems/edit/1297092

The issue was on char type when randomly the length of the char in the test was set a large value (example: > 3K).

**Fix**
The problem was with how the caller was calling IsValidUTF8 and its signature.
We've been sending `char* str` in the caller but receiving `string& str` in the function signature. The compiler wasn't complaining but every time we have called this function, the value of string had to allocated/deleted on the stack which was time-consuming when the number of calls/ length of strings was large. This change normally calls the function by reference.

**Test:**
- Unit tests were passed locally.
- Testshell tests were passed locally.

PS: a small change comes with this PR when `-j` was added to Windows build for enabling parallel build.

Gonna get the nuget and test with PVS on DS Main Dev.