Merged PR 4901442: Upgrading R version in build pipeline
**Issue**
Using the generated R-lang-extension Nuget as of now raises the following error:
Error: function 'Rcpp_precious_remove' not provided by package 'Rcpp'

Also in the build pipeline, we can see a couple of warnings related to this:
![image.png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/4901442/attachments/image.png)

![image (2).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/4901442/attachments/image%20%282%29.png)

**Fix**
In the build pipeline, we install packages Rcpp/RInside by downloading them from a cran repository. Rcpp/RInside current version is built by 4.0.5 but our pipeline builds RExtension with R.4.0.3. We should upgrade the R version that is used in the pipeline.
To prevent happening this next time, the version of Rcpp/RInside got fixed to the existing latest version of them in cran archive.

**Tests**
All unit tests were passed in the pipeline.
Without this change, testshell tests are failing with the new Nuget.