Merged PR 4901360: [.NET Core C# Extension] Package the CSharpExtensionTest library for use in TestShellTest
[**Issue**]
In order to add E2E C# Extension tests, we need a user library with classes that implement the AbstractSqlServerExtensionExecutor for execution of custom scripts. We build such a test library in this repo but dont export it currently.

[**Fix**]
Although we could write similar code implementing the user libraries in TestShell, to avoid duplicate code this fix exports the test library so it will be packaged in the same nuget that we already consume for the extension itself.

An additional minor fix to build script to error out on failure
[**Test**]
Validated the YAML is correct online.