Merged PR 5866453: Update pipeline to be built on ubuntu20.04 container
### Background
From SQL Server 2019 CU10 onwards, mssql-server-* packages are shipped only for Ubuntu 20.04
See the built packages here: http://helsinki/dist/packages/ubuntu2004/

From this extensions repository, we only consume the SDK `mssql-java-lang-extension.jar` and publish it here: http://helsinki/dist/packages/java/extensibility/java-lang-extension/releases
for e.g. the last version published:
http://helsinki/dist/packages/java/extensibility/java-lang-extension/releases/1.0.0-CI-20190507-042302/
Hence, we need to get the latest version of the sdk `mssql-java-lang-extension.jar` to ship on Ubuntu20.04 package

### This change
- Specify the 20.04 container in .yml
- Fix `restore-packages.sh` scripts to make them work on ubuntu20.04
- Unfortunately, with this update, the python unit tests fail to initialize numpy even though it has been installed. Since we dont ship python extension, not running the tests on linux for now.

### Note
The `javaextension.so` that we build here is NOT shipped as part of the `mssql-server-extensibility-java` debian package.
The `mssql-server-extensibility-java` debian package is made up of:
- `mssql-java-lang-extension.jar` from this repo and
- `javaextension.so` that is built in `mssql-server` repo.