Merged PR 6964590: Python Language Extension varchar(max) and nvarchar(max) truncation bugfix
# Why make this change?
Bugfix for https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/1972773

# Background
Python Extension will truncate the output of a string parameter to `USHRT_MAX`, which is 65535, in the case where we have `varchar(max)` or `nvarchar(max)`. This is because we truncate the output to be the size of the parameter that was initialized in `InitParam`. `ExtHost` passes in this size as `SQLSatellite_TypeInfo->m_lengthInByte`, which for these cases is `USHRT_MAX`, which is 65535. The desired behavior, however, would be for `varchar(max)` and `nvarchar(max)` to not be truncated in the output.

# What is this change?
This is fixed by modifying the function `PythonStringParam<CharType>::RetrieveValueAndStrLenInd` from within `PythonParam.cpp` to check if the size of the parameter is equal to `USHRT_MAX`, and in that case to not truncate the output.
![image.png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/6964590/attachments/image.png)

# How was this tested?
Testing was done manually in SSMS using the following query for both `varchar(max)` and `nvarchar(max)`

![image (3).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/6964590/attachments/image%20%283%29.png)

We also, in `PythonGetOutputParamTests.cpp`, modify the tests `GetStringOutputParamTest` and `GetWStringOutputParamTest` to add parameters that will be initialized in `InitParam` with a size of `USHRT_MAX`, but that actually contain values that exceed `USHRT_MAX` in size. The tests then validate that these parameters are not truncated by Python Extension.