Merged PR 7025705: Add Linux PR pipeline with updated image
## Why the change is needed
We have official windows, official linux, and windows pr pipelines, but the Linux PR pipeline is missing. Also, due to the docker image update for windows pipelines, there are some build errors of linux official pipeline caused by the experimental::filesystem migrating to filesystem. The linux image does not support <filesystem> yet.
## What is the change
This PR adds a new Linux PR pipeline and updates both Linux Official ubuntu image from 18.04 to 20.04. The PR adds a check for the current OS when we need filesystem. If it is linux we import experimental::filesystem instead. The PR also has a small fix calling empty-script.sh in the pipeline.