Merged PR 7203126: Revert 'R Language Extension Truncation Bugfix for varchar(max), nvarchar(max), and varbinary(max)'
# Reverting change due to seg fault in Linux environment
In a Linux environment when we call `print` from within `R` on sufficiently large `wide string` variables we encounter a segmentation fault.  We revert the change described below until a fix for this is available.

# Why make this change?
Bugfix for `R Language Extension` truncating the characters of an Output parameter of type `varchar(max)`, `nvarchar(max)`, or `varbinary(max)` to `65535` without any modifications in the `R` script. We also disover a seg fault in our test environment when using `message(param)` within `R` when `param` is sufficiently large (over 63451). `message(param)` is therefore replaced with `print(param)` to avoid the segmentation fault.

# Background
R Extension will truncate the output of a parameter to `USHRT_MAX`, which is `65535`, in the case where we have `varchar(max)`, `nvarchar(max)`, or `varbinary(max)`. This is because we truncate the output to be the size of the parameter that was initialized. `ExtHost` passes in this size as `SQLSatellite_TypeInfo->m_lengthInByte`, which for these cases is `USHRT_MAX`, which is `65535`. The desired behavior, however, would be for `varchar(max)`, `nvarchar(max)`, and `varbinary(max)` to not be truncated in the output.

# What is this change?
![image.png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/7130733/attachments/image.png)
![image (2).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/7130733/attachments/image%20%282%29.png)
This is fixed by modifying the function `RTypeUtils::FillDataFromCharacterVector` and `RTypeUtils::FillDataFromRawVector` from within `RTypeUtils.cpp` to check if the size of the parameter is equal or greater than `USHRT_MAX`, and in that case to not truncate the output.

Lastly, we modify the function `Logger::LogRVariable(const string &name)` to use `print` instead of `message`.

![image (4).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/7130733/attachments/image%20%284%29.png)
# How was this tested?
Testing was done manually in SSMS using a query of the following general form for varchar(max), nvarchar(max), and varbinary(max)
![image (3).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/7130733/attachments/image%20%283%29.png)

We also, in `RExtensionGetOutputParamTests.cpp`, modify the tests `GetCharOutputParamTest`, `GetRawOutputParamTest`, and `GetNCharOutputParamTest` to add parameters that will be initialized with a size greater than `USHRT_MAX`. The tests then validate that these parameters are not truncated by `RExtension`.

Reverts !7130733