Merged PR 7474801: [R Extension] Removing dependency on cv2pdb package
Since we have a new nuget feed for Data-sql-language-extension, we need to clean up the stale packages.
Right now, the new feed depends on the deprecated feed due to the dependency of cv2pdb package.
Also, as part of the secure supply chain requirements, we should not use tools that don't have a secure lineage.
cv2pdb is an open-source tool but there is no support for a secure build for this.
Therefore, this PR removes the dependency of cv2pdb.
Work item: https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/2100331/