Merged PR 7619671: Fixing java extension Linux build pipeline by removing an extra package installation step
**Issue**
Recently, the build pipeline for the java extension on Linux started to fail with the following error:
`/usr/include/sqltypes.h:56:10: fatal error: unixodbc.h: No such file or directory`

**RCA**
The error is related to version `2.3.11` of the package `unixodbc-dev` in which the recent version needs to #include "unixodbc.h" in its `sqltypes.h` which cannot be found on ubuntu 20.04 even after installation of the following packages:
libodbc1
unixodbc-dev
unixodbc
odbcinst1debian2
odbcinst

In the build pipeline, we install `unixodbc-dev` package in the generic [`restore-packages.sh`](https://dev.azure.com/msazure/One/_git/Data-SQL-Language-Extensions?path=/restore-packages.sh) and also in the Java extension [one](https://dev.azure.com/msazure/One/_git/Data-SQL-Language-Extensions?path=/language-extensions/java/build/linux/restore-packages.sh) (`language-extensions/java/build/linux/restore-packages.sh`). The one that is getting installed in the generic restore packages is version `3.7` which does not have compatibility issues (the installation there is done before performing the `apt-get` update).

**Fix**
This PR just removes the installation of the package in the java extension `restore-packages.sh`. We already install  `unixodbc-dev` in the previous steps of the build pipeline.

An alternative way for this was to fix the version of installation to `3.7` until we resolve the compatibility issues with the recent version of `unixodbc-dev`;
`apt-get install unixodbc=2.3.7`

This is needed to unblock the pipeline as soon as possible.
I've created a task to not lose the context and follow up on upgrading it to the latest version.
https://msdata.visualstudio.com/Database%20Systems/_workitems/edit/2232485