Merged PR 7625566: Updating the windows container image for the build pipeline
**RE: Action required from build pipeline's owners using cdpxwin1809.azurecr.io/global/obinfra/windows/1809/vs2019:latest**

The Onebranch team is making an update to the cached containers on the build agent systems. The reason for this change is to maintain the highest security and compliance standards for our infrastructure.

As a result, the container we use here to be pulled from the build agent cache on February 21st, 2023.  After the container is removed from the cache on Feb 21st, customers using this container will experience a significant delay in container initialization times of up to an additional hour.

To avoid this delay, it is recommended to move to one of their latest fully supported global containers.
For more details on recommended containers to be using see this [wiki](https://onebranch.visualstudio.com/OneBranch/_wiki/wikis/OneBranch.wiki/4587/Container-Images)

This PR changes the old container image `cdpxwin1809.azurecr.io/global/vse2019:latest`  to `onebranch.azurecr.io/windows/ltsc2019/vse2019:latest`; Note that even though the path of the old container is a bit different from what's mentioned in the email but going forward `onebranch.azurecr.io/windows/ltsc2019/vse2019:latest` is the latest global windows container to be used anyway. So, to be safe, this change has been done in this PR.