Merged PR 7473832: R Language Extension Truncation bug fix with segmentation fault fix
# Description
## Why make this change?
Bugfix for R Language Extension truncating the characters of an Output parameter of type varchar(max), nvarchar(max), or varbinary(max) to 65535 without any modifications in the R script. We also discover a seg fault in our test environment when using message(param) within R when param is sufficiently large (over 63451). message(param) is therefore replaced with print(param) to avoid the segmentation fault. However, `for the R runtime` this continued to be a problem in the Linux environment. Further investigation discovered that the multibyte wide strings in the Linux environment had to be created carefully from vectors and arrays, with attention to making sure that the null terminator was placed correctly. However, even with this careful creation, values large enough to test the truncation bug would still result in segmentation fault, caused by a double free later in the RAW binaries test when using siufficiently large values for 'NVARCHAR'. For this reason, we do not test large value for `NVARCHAR`.

![image (4).png](https://dev.azure.com/msazure/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/7473832/attachments/image%20%284%29.png)

## Background
R Extension will truncate the output of a parameter to USHRT_MAX, which is 65535, in the case where we have varchar(max), nvarchar(max), or varbinary(max). This is because we truncate the output to be the size of the parameter that was initialized. ExtHost passes in this size as SQLSatellite_TypeInfo->m_lengthInByte, which for these cases is USHRT_MAX, which is 65535. The desired behavior, however, would be for varchar(max), nvarchar(max), and varbinary(max) to not be truncated in the output.

## What is this change?

This is fixed by modifying the function RTypeUtils::FillDataFromCharacterVector and RTypeUtils::FillDataFromRawVector from within RTypeUtils.cpp to check if the size of the parameter is equal or greater than USHRT_MAX, and in that case to not truncate the output. We do not test `NVARCHAR` due to it causing segmentation fault in the pipeline.

Lastly, we modify the function Logger::LogRVariable(const string &name) to use print instead of message.

How was this tested?
Testing was done manually in SSMS using a query of the following general form for varchar(max), nvarchar(max), and varbinary(max)

```
CREATE OR ALTER PROCEDURE dbo.P_TestR
AS
BEGIN
DECLARE @RequestStr VARCHAR(max)=replicate('A',10000),
@ResponseStr VARCHAR(max),
@i INT=1

WHILE @i<=4 BEGIN
    SELECT @RequestStr=@RequestStr+RequestStr
    SET @i+=1
END
-- Prints 128000
select len(@RequestStr)

EXEC sp_execute_external_script
    @language=N'myR',
    @script=N'
print(len(RequestStr))
ResponseStr = RequestStr
',
@params=N'@RequestStr VARCHAR(MAX),@ResponseStr VARCHAR(MAX) OUTPUT
@RequestStr=@RequestStr,
@ResponseStr=@ResponseStr OUTPUT

-- Prints 128000
select len(@ResponseStr)
END
go
EXEC dbo.P_TestR
go
-- Output 128000,1...