Merged PR 7713901: [Java Extension] Modify JVM library relative path sourced from JRE_HOME to JAVA_HOME
Starting with Java 11, there is no separate JRE anymore. Only the JDK is offered, which includes a runtime environment. The current relative path to the JVM library is still sourced from the JRE_HOME which we do not ask user to provide anymore.
Therefore, the change makes JVM library to be found according to the JAVA_HOME.