Merged PR 7774162: Initialize git submodule with modified pipelines
**What is the change**
This PR deletes all source code and add a git submodule referencing the open source repo: https://github.com/microsoft/sql-server-language-extensions.
The pipelines are modified to build projects from the submodule.

**Remark**
Currently since the restore-package.cmd for python is not open sourced yet, the **restore-python-packages.cmd** is added temporarily to restore the packages for python during pipeline. Once the restore-packages.cmd for python is ready at the open-source repo, it will be removed along with the **packages.python.config**.