Merged PR 9484155: [Linux pipeline] Fix linux pipeline to output java extension correctly
This PR addresses a critical issue where the Linux output from our build pipeline was missing the Java extension (`libJavaExtension.so.1.0`) in the `drop/build/out` folder. This omission resulted in the NuGet package lacking the necessary Java extension. However this is not a prod issue since the released java extension in the OSS is the zip file containing the java extension as explained in detail below.

The root of this issue was traced to the existence of the Java extension in two distinct locations within our build output: directly within the `drop/build/out` folder and within a zip file located at `drop/package/out/java-extension/` (**see attached screenshot**). Although the zip file intended for customer releases contained the extension, making this discrepancy initially non-blocking for release, it was nonetheless incorrect and required rectification to ensure consistency and correctness across our build artifacts. (for example, the java extension for linux isn't included in the nuget we use for dsmaindev/mssql-server testing.)
![image (2).png](https://msazure.visualstudio.com/b32aa71e-8ed2-41b2-9d77-5bc261222004/_apis/git/repositories/6376a113-07ad-4082-bd2b-7d8aaf437df4/pullRequests/9484155/attachments/image%20%282%29.png)

### Fix Implemented:

This update corrects the Java extension's name and path within the build pipeline configuration to ensure it is consistently included in the correct location. The adjustments made are as follows:

**Previous Configuration:**

- **Contents Path:** `**/libjavaextension.so.1.0`
- **Target Folder:** `$(Build.SourcesDirectory)\out\build\out\java-extension`

**Updated Configuration:**

- **Contents Path:** `**/libJavaExtension.so.1.0`
- **Target Folder:** `$(Build.SourcesDirectory)/out/build/out/java-extension`

By aligning the Java extension's naming convention and ensuring its presence in the expected directory, this change guarantees that all Linux build outputs will include `libJavaExtension.so.1.0` as expected, thereby preventing any future discrepancies in NuGet package contents and enhancing the reliability of our release process.