Merged PR 13337258: [Data-SQL-Language-Extensions] Pipeline Failure Fixes
## Why is this change being made?
We have had recent failures in the build pipelines for PRs and Official builds for this repository. This is a blocker for getting code changes in, releasing new builds and keeping the repository healthy. We need to ensure that the pipelines are healthy to unblock ongoing work and ensure that new releases from the pipeline are possible.

## What does this change do?
This change has a second part made entirely in the OSS repository and consumed here via the git submodule - https://github.com/microsoft/sql-server-language-extensions/pull/62

- `UsePythonVersion@0` devops task is now used to provide python for the build purpose instead of downloading python in the `restore-packages.cmd` script.
- The installation path from `UsePythonVersion@0` task is passed into the build scripts and reused during building.
- Python version remains the same as was previously used - `3.10.2`
- Updates `PublishCodeCoverageResults` from v1 to v2 as v1 is now deprecated and now generates a warning during the build.
- `dotnet restore` for the CSharpExtension project is now done via the devops task `DotNetCoreCLI@2` instead of via the `restore-packages.cmd` script as a mitigation to the current failure of the `dotnet restore` command from the batch script.

## How is this change tested?
Scheduled pipeline runs were successful.

## Associated work item
https://dev.azure.com/msdata/Database%20Systems/_workitems/edit/4574724/