Add static nethost library from .NET 8.0 Linux runtime.
We will use the nethost to locate the system's installed hostfxr, as
using a local libhostfxr.so breaks the .NET runtime lookup.

nethost contains the logic needed to find the .NET root based on the
DOTNET_ROOT environment variable or the install_location configuration
file, which we will require for .NET runtime lookup on diverging Linux
distributions (e.g. RHEL and Debian packages not installing .NET in the
same base directories).